--- Front-end/src/app/(pages)/assistant/Chat/chatbot.module.scss
@@ -1,203 +1,0 @@
-.chat-container {
-    display: flex;
-    flex-direction: column;
-    height: 100% !important;
-
-    .message-container {
-        font-size: 4rem;
-        flex-grow: 1;
-        font-weight: 500;
-        overflow-y: auto;
-        margin-bottom: 1rem;
-    }
-
-    .chat-message {
-        display: flex;
-        flex-direction: column;
-        margin-bottom: 1rem;
-
-        &.message-start {
-            align-items: flex-start;
-        }
-
-        &.message-end {
-            align-items: flex-end;
-        }
-    }
-
-    .avatar-container {
-        margin: 0.5rem;
-
-        .avatar-wrapper {
-            width: 2.5rem;
-            height: 2.5rem;
-            border-radius: 50%;
-            overflow: hidden;
-
-            @media (min-width: 768px) {
-                width: 3rem;
-                height: 3rem;
-            }
-
-            .avatar-image {
-                width: 100%;
-                height: 100%;
-                object-fit: cover;
-            }
-        }
-    }
-
-    .chat-content {
-        height: 100% !important;
-    }
-
-    .message-bubble {
-        background-color: #f3f4f6;
-        padding: 0.75rem 1rem;
-        border-radius: 1rem;
-        max-width: 80%;
-
-        &.model-bubble {
-            background-color: #3b82f6;
-            color: white;
-        }
-    }
-
-        .input-container {
-            height: auto;
-            display: flex;
-            align-items: center;
-            /* Căn giữa nội dung theo chiều dọc */
-            gap: 0.5rem;
-            padding: 0.5rem;
-            background-color: #f9fafb;
-            border-top: 1px solid #e5e7eb;
-        }
-    
-        .chat-input {
-            flex-grow: 1;
-            height: 2.5rem;
-            line-height: 1.5;
-            overflow: hidden;
-            resize: none;
-            padding: 0.5rem 0.75rem;
-            border-radius: 0.5rem;
-            border: 1px solid #e5e7eb;
-            background-color: white;
-            transition: all 0.2s;
-    
-            &:focus {
-                outline: none;
-                border-color: #3b82f6;
-                box-shadow: 0 0 0 2px rgba(59, 130, 246, 0.2);
-            }
-    
-            &::-webkit-scrollbar {
-                display: none;
-                /* Ẩn thanh cuộn */
-            }
-        }
-
-    .reset-button,
-    .send-button {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        padding: 0.75rem;
-        border-radius: 0.5rem;
-        transition: all 0.2s;
-
-        &:disabled {
-            opacity: 0.5;
-            cursor: not-allowed;
-        }
-    }
-
-    .reset-button {
-        border: 1px solid #ef4444;
-        color: #ef4444;
-
-        &:hover:not(:disabled) {
-            background-color: #ef4444;
-            color: white;
-        }
-    }
-
-    .send-button {
-        background-color: #3b82f6;
-        color: white;
-
-        &:hover:not(:disabled) {
-            background-color: #2563eb;
-        }
-
-        &.loading {
-            background-color: #8b5cf6;
-        }
-    }
-
-    .icon-small {
-        width: 1rem;
-        height: 1rem;
-    }
-
-    .spinner {
-        width: 1rem;
-        height: 1rem;
-        border: 2px solid transparent;
-        border-top-color: currentColor;
-        border-radius: 50%;
-        animation: spin 1s linear infinite;
-    }
-}
-
-@keyframes spin {
-    to {
-        transform: rotate(360deg);
-    }
-}
-
-
-// toogle
-.chat-toggle-button {
-    position: fixed;
-    bottom: 1rem;
-    right: 1rem;
-    width: 3rem;
-    height: 3rem;
-    background-color: #3b82f6;
-    border: none;
-    border-radius: 50%;
-    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.2);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    cursor: pointer;
-    transition: all 0.3s;
-
-    &:hover {
-        background-color: #2563eb;
-    }
-
-    svg {
-        width: 1.5rem;
-        height: 1.5rem;
-    }
-}
-
-.chat-container {
-    font-size: 1.5rem;
-    position: fixed;
-    max-height: 400px;
-    color: black;
-    bottom: 5rem;
-    right: 1rem;
-    width: 90%;
-    max-width: 350px;
-    background-color: white;
-    border: 1px solid #e5e7eb;
-    border-radius: 0.5rem;
-    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
-    overflow: hidden;
-    z-index: 1000;
-}
