+++ Front-end/src/app/(pages)/chat/Chat.module.scss
@@ -1,1 +1,201 @@
-@import "../../../styles/mixin.scss";
+.wrapper {
+    display: flex;
+    flex-direction: column;
+    min-height: 100vh;
+    background-color: #2d2d2d; // Dark gray background
+}
+
+.content {
+    flex: 1;
+    width: 100%;
+    max-width: 1200px;
+    margin: 0 auto;
+    padding: 0 1rem;
+}
+
+.chat-container {
+    display: flex;
+    flex-direction: column;
+    height: calc(100vh - 140px);
+    margin: 0 auto;
+    max-width: 800px;
+    width: 100%;
+    position: relative;
+}
+
+.message-container {
+    flex: 1;
+    overflow-y: auto;
+    padding: 2rem 0;
+    scroll-behavior: smooth;
+    // background-color: #2d2d2d; // Dark gray background
+
+    &::-webkit-scrollbar {
+        width: 6px;
+    }
+
+    &::-webkit-scrollbar-track {
+        background: #404040;
+    }
+
+    &::-webkit-scrollbar-thumb {
+        background: #666;
+        border-radius: 3px;
+    }
+}
+
+.chat-message {
+    display: flex;
+    padding: 1.5rem 1rem;
+    gap: 1rem;
+    color: #ff69b4;
+    font-size: 1.25rem;
+
+    &:hover {
+        background-color: #404040;
+    }
+}
+
+
+.avatar-container {
+    flex-shrink: 0;
+}
+
+.avatar-wrapper {
+    width: 50px;
+    height: 50px;
+    border-radius: 50%;
+    overflow: hidden;
+    border: 2px solid #ff69b4; // Pink border for avatars
+}
+
+.avatar-image {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+}
+
+.message-bubble {
+    flex: 1;
+    font-size: 1.25rem;
+    line-height: 1.6;
+    white-space: pre-wrap;
+    overflow-wrap: break-word;
+    color: #ff69b4;
+    background-color: #454545; // Slightly lighter gray for message bubbles
+    padding: 1rem;
+    border-radius: 8px;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+
+    p {
+        margin: 0.5rem 0;
+
+        &:first-child {
+            margin-top: 0;
+        }
+
+        &:last-child {
+            margin-bottom: 0;
+        }
+    }
+
+    code {
+        background-color: #333;
+        padding: 0.2em 0.4em;
+        border-radius: 3px;
+        font-family: monospace;
+        color: #ff69b4;
+    }
+
+    pre {
+        background-color: #333;
+        padding: 1rem;
+        border-radius: 6px;
+        overflow-x: auto;
+        margin: 1rem 0;
+        font-size: 1.2rem;
+    }
+}
+
+.model-bubble {
+    color: #ff69b4;
+}
+
+.input-container {
+    position: sticky;
+    bottom: 0;
+    width: 100%;
+    padding: 1rem;
+    background-color: #2d2d2d; // Dark gray background
+    border-top: 1px solid #404040;
+    display: flex;
+    gap: 0.5rem;
+    align-items: flex-end;
+}
+
+.chat-input {
+    flex: 1;
+    max-height: 200px;
+    padding: 10px;
+    font-size: 1.25rem;
+    line-height: 1.5;
+    border: 1px solid #404040;
+    border-radius: 0.75rem;
+    resize: none;
+    outline: none;
+    background-color: #3d3d3d; // Lighter gray for input
+    color: #ff69b4;
+
+    &:focus {
+        border-color: #ff69b4;
+        box-shadow: 0 0 0 2px rgba(255, 105, 180, 0.1);
+    }
+
+    &::placeholder {
+        color: rgba(255, 105, 180, 0.6);
+    }
+}
+
+.send-button,
+.reset-button {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 40px;
+    height: 40px;
+    border-radius: 0.5rem;
+    border: none;
+    background-color: #3d3d3d; // Lighter gray for buttons
+    cursor: pointer;
+    transition: background-color 0.2s;
+
+    &:hover {
+        background-color: #4d4d4d;
+    }
+
+    &:disabled {
+        cursor: not-allowed;
+        opacity: 0.5;
+    }
+}
+
+.icon-small {
+    width: 20px;
+    height: 20px;
+    color: #ff69b4;
+}
+
+.spinner {
+    width: 20px;
+    height: 20px;
+    border: 2px solid #3d3d3d;
+    border-top-color: #ff69b4;
+    border-radius: 50%;
+    animation: spin 1s linear infinite;
+}
+
+@keyframes spin {
+    to {
+        transform: rotate(360deg);
+    }
+}
